Rename initcontracts to deploy
