+++ command/rootchain/deploy/deploy_test.go
@@ -1,4 +1,4 @@
-package initcontracts
+package deploy
 
 import (
 	"os"
