+++ command/rootchain/rootchain.go
@@ -3,8 +3,8 @@ package rootchain
 import (
 	"github.com/spf13/cobra"
 
+	"github.com/0xPolygon/polygon-edge/command/rootchain/deploy"
 	"github.com/0xPolygon/polygon-edge/command/rootchain/fund"
-	"github.com/0xPolygon/polygon-edge/command/rootchain/initcontracts"
 	"github.com/0xPolygon/polygon-edge/command/rootchain/server"
 )
 
@@ -19,7 +19,7 @@ func GetCommand() *cobra.Command {
 		// rootchain server
 		server.GetCommand(),
 		// rootchain deploy
-		initcontracts.GetCommand(),
+		deploy.GetCommand(),
 		// rootchain fund
 		fund.GetCommand(),
 	)
